[New] Added test cases for record form form subforms subform field array
