+++ app/javascript/components/record-form/form/subforms/subform-field-array/component.jsx
@@ -105,10 +105,10 @@ const Component = ({
   );
 
   return (
-    <div className={css.fieldArray}>
+    <div className={css.fieldArray} data-testid="subForm-field-array">
       <div className={cssContainer}>
         {!renderAsAccordion && (
-          <div>
+          <div data-testid="subForm-header">
             <h3 className={css.subformTitle}>
               {renderAddFieldTitle} {title} {parentTitle}
             </h3>
