+++ app/javascript/components/record-form/form/subforms/subform-fields/component.jsx
@@ -145,6 +145,7 @@ const Component = ({
 
           return (
             <ListItem
+              data-testid="list-item"
               onClick={handleEdit(index)}
               component="a"
               classes={{ divider: css.listDivider, root: css.listItem }}
